requirements.txt
# Core scientific stack
numpy>=1.26,<2.1
scipy>=1.11,<2.0
pandas>=2.1,<3.0
matplotlib>=3.8,<4.0

# Data & IO
h5py>=3.10,<4.0
tables>=3.9,<4.0
imageio>=2.34
pillow>=10.2,<11.0
opencv-python>=4.9

# ML / DL core
# torch==2.3.1
# torchvision==0.18.1
# torchaudio==2.3.1
# triton>=2.3,<3.0
# xformers>=0.0.26,<0.0.28
timm>=0.9.12

# Utilities
einops>=0.7
tqdm>=4.66
tabulate>=0.9
termcolor>=2.4

# Geometry / 3D
trimesh>=4.3

# Experiment tracking
wandb>=0.16

# Hugging Face ecosystem
huggingface-hub>=0.22,<1.0

# Visualization / UI
gradio>=4.0

# Serialization / protocol
protobuf>=4.25,<5.0